Moved database files into database directory
