+++ app/src/main/java/edu/cascadia/mobas/pollinatorpathway/ui/Database/BoxDao.java
@@ -1,4 +1,4 @@
-package edu.cascadia.mobas.pollinatorpathway.ui.login;
+package edu.cascadia.mobas.pollinatorpathway.ui.Database;
 
 import androidx.room.Dao;
 import androidx.room.Delete;
@@ -6,7 +6,6 @@ import androidx.room.Insert;
 import androidx.room.OnConflictStrategy;
 import androidx.room.Query;
 import androidx.room.Update;
-import java.util.List;
 
 
 @Dao
