+++ app/src/main/java/edu/cascadia/mobas/pollinatorpathway/ui/Database/Planting.java
@@ -1,4 +1,4 @@
-package edu.cascadia.mobas.pollinatorpathway.ui.login;
+package edu.cascadia.mobas.pollinatorpathway.ui.Database;
 
 import static androidx.room.ForeignKey.CASCADE;
