+++ app/src/main/java/edu/cascadia/mobas/pollinatorpathway/ui/Database/ProfileDao.java
@@ -1,4 +1,4 @@
-package edu.cascadia.mobas.pollinatorpathway.ui.login;
+package edu.cascadia.mobas.pollinatorpathway.ui.Database;
 
 import androidx.room.Dao;
 import androidx.room.Delete;
