+++ app/src/main/java/edu/cascadia/mobas/pollinatorpathway/ui/Database/UserDatabase.java
@@ -1,4 +1,4 @@
-package edu.cascadia.mobas.pollinatorpathway.ui.login;
+package edu.cascadia.mobas.pollinatorpathway.ui.Database;
 
 import android.content.Context;
